+++ .gitignore
@@ -24,3 +24,4 @@ CognitiveServices.js
 CognitiveServices.js.map
 *.css
 /test/uploads
+.env
