+++ package.json
@@ -16,6 +16,7 @@
     "webpack": "webpack",
     "webpack-watch": "webpack --progress --watch",
     "clean": "rimraf built",
+    "postinstall": "cd test && npm install",
     "prepublish": "tsc && npm run build-css && webpack --progress --config webpack.production.config.js",
     "start": "http-server -p 8000",
     "test": "concurrently  \"node test/mock_dl/index.js\" \"mocha test\" ",
@@ -43,29 +44,23 @@
     "tslib": "1.7.1"
   },
   "devDependencies": {
-    "@types/body-parser": "0.0.33",
     "@types/chai": "3.4.34",
     "@types/chai-subset": "1.0.29",
     "@types/deep-freeze": "0.0.29",
-    "@types/express": "4.0.35",
     "@types/markdown-it": "0.0.2",
     "@types/nightmare": "1.6.30",
     "@types/react-dom": "0.14.23",
     "@types/react-redux": "4.4.38",
     "awesome-typescript-loader": "3.1.2",
-    "body-parser": "1.16.0",
     "chai": "3.5",
     "chai-subset": "1.4.0",
     "concurrently": "3.1.0",
     "deep-freeze": "0.0.1",
-    "dotenv": "4.0.0",
     "expose-loader": "0.7.3",
-    "express": "4.14.0",
     "fs-extra": "1.0.0",
     "http-server": "0.9.0",
     "microsoft-adaptivecards-scss-to-json": "1.2.0",
     "mocha": "3.2.0",
-    "multer": "1.3.0",
     "nightmare": "2.8.1",
     "nightmare-upload": "0.1.1",
     "nightmare-window-manager": "^0.2.2",
