+++ test/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "botframework-webchat-mockdl",
+  "version": "0.0.1",
+  "description": "> ## Disclaimer: This test framework is intended for Web Chat and DirectLine testing ONLY. This is NOT intended for testing Bots.",
+  "main": "mock_dl/index.js",
+  "scripts": {
+    "postinstall": "tsc -p ./mock_dl && tsc -p . && tsc -p ./mock_speech",
+    "start": "node mock_dl"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git://git@github.com/Microsoft/BotFramework-WebChat.git"
+  },
+  "author": "Microsoft Corp",
+  "license": "MIT",
+  "dependencies": {
+    "@types/body-parser": "^1.16.8",
+    "@types/express": "^4.0.39",
+    "@types/nightmare": "^2.10.0",
+    "@types/node": "^8.0.56",
+    "body-parser": "^1.18.2",
+    "botframework-directlinejs": "0.9.13",
+    "dotenv": "^4.0.0",
+    "express": "^4.16.2",
+    "multer": "^1.3.0",
+    "nightmare": "^2.10.0",
+    "typescript": "2.3"
+  }
+}
